requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
gunicorn==21.2.0
python-dotenv==1.0.0
firebase-admin==6.5.0
pillow==11.3.0
python-multipart==0.0.6
